+++ api/internal/domain/subexpression.go
@@ -10,7 +10,7 @@ type SubExpression struct {
 	SubExpressionId1  *int64
 	SubExpressionId2  *int64
 	Operation         string
-	OperationDuration uint16
+	OperationDuration uint32
 	Result            *float64
 	Agent             string
 	IsLast            bool
@@ -22,6 +22,6 @@ type ReadySubExpression struct {
 	Val1      float64 `json:"val1"`
 	Val2      float64 `json:"val2"`
 	Operation string  `json:"operation"`
-	Duration  uint16  `json:"operation_duration"`
+	Duration  uint32  `json:"operation_duration"`
 	IsLast    bool    `json:"is_last"`
 }
